UART Device Driver Porting
