+++ dts/arm/tcc/tcc70xx.dtsi
@@ -100,7 +100,7 @@
 	clocks {
 		clk_uart: clk_uart {
 			compatible = "fixed-clock";
-			clock-frequency = <200000000>;
+			clock-frequency = <48000000>;
 			#clock-cells = <0>;
 		};
 	};
